+++ crates/bevy_material/src/render/mod.rs
@@ -1,7 +1,7 @@
 mod mesh;
 // mod mesh_bindings;
-// mod mesh_view_bindings;
+mod mesh_view_bindings;
 
 pub use mesh::*;
 // pub use mesh_bindings::*;
-// pub use mesh_view_bindings::*;
+pub use mesh_view_bindings::*;
