+++ crates/bevy_pbr/src/render/mesh_view_bindings.rs
@@ -7,17 +7,16 @@ use bevy_core_pipeline::{
         get_lut_bind_group_layout_entries, get_lut_bindings, Tonemapping, TonemappingLuts,
     },
 };
-use bevy_derive::{Deref, DerefMut};
 use bevy_ecs::{
     component::Component,
     entity::Entity,
     query::Has,
-    resource::Resource,
     system::{Commands, Query, Res},
     world::World,
 };
 use bevy_image::BevyDefault as _;
 use bevy_light::{EnvironmentMapLight, IrradianceVolume};
+use bevy_material::render::{MeshPipelineViewLayout, MeshPipelineViewLayoutKey, MeshPipelineViewLayouts};
 use bevy_math::Vec4;
 use bevy_render::{
     globals::{GlobalsBuffer, GlobalsUniform},
@@ -45,115 +44,13 @@ use crate::{
     },
     prepass, EnvironmentMapUniformBuffer, FogMeta, GlobalClusterableObjectMeta,
     GpuClusterableObjects, GpuFog, GpuLights, LightMeta, LightProbesBuffer, LightProbesUniform,
-    MeshPipeline, MeshPipelineKey, RenderViewLightProbes, ScreenSpaceAmbientOcclusionResources,
+    MeshPipeline, RenderViewLightProbes, ScreenSpaceAmbientOcclusionResources,
     ScreenSpaceReflectionsBuffer, ScreenSpaceReflectionsUniform, ShadowSamplers,
     ViewClusterBindings, ViewShadowBindings, CLUSTERED_FORWARD_STORAGE_BUFFER_COUNT,
 };
 
 #[cfg(all(feature = "webgl", target_arch = "wasm32", not(feature = "webgpu")))]
 use bevy_render::render_resource::binding_types::texture_cube;
-
-#[cfg(debug_assertions)]
-use {crate::MESH_PIPELINE_VIEW_LAYOUT_SAFE_MAX_TEXTURES, bevy_utils::once, tracing::warn};
-
-#[derive(Clone)]
-pub struct MeshPipelineViewLayout {
-    pub main_layout: BindGroupLayoutDescriptor,
-    pub binding_array_layout: BindGroupLayoutDescriptor,
-    pub empty_layout: BindGroupLayoutDescriptor,
-
-    #[cfg(debug_assertions)]
-    pub texture_count: usize,
-}
-
-bitflags::bitflags! {
-    /// A key that uniquely identifies a [`MeshPipelineViewLayout`].
-    ///
-    /// Used to generate all possible layouts for the mesh pipeline in [`generate_view_layouts`],
-    /// so special care must be taken to not add too many flags, as the number of possible layouts
-    /// will grow exponentially.
-    #[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
-    #[repr(transparent)]
-    pub struct MeshPipelineViewLayoutKey: u32 {
-        const MULTISAMPLED                = 1 << 0;
-        const DEPTH_PREPASS               = 1 << 1;
-        const NORMAL_PREPASS              = 1 << 2;
-        const MOTION_VECTOR_PREPASS       = 1 << 3;
-        const DEFERRED_PREPASS            = 1 << 4;
-        const OIT_ENABLED                 = 1 << 5;
-    }
-}
-
-impl MeshPipelineViewLayoutKey {
-    // The number of possible layouts
-    pub const COUNT: usize = Self::all().bits() as usize + 1;
-
-    /// Builds a unique label for each layout based on the flags
-    pub fn label(&self) -> String {
-        use MeshPipelineViewLayoutKey as Key;
-
-        format!(
-            "mesh_view_layout{}{}{}{}{}{}",
-            if self.contains(Key::MULTISAMPLED) {
-                "_multisampled"
-            } else {
-                Default::default()
-            },
-            if self.contains(Key::DEPTH_PREPASS) {
-                "_depth"
-            } else {
-                Default::default()
-            },
-            if self.contains(Key::NORMAL_PREPASS) {
-                "_normal"
-            } else {
-                Default::default()
-            },
-            if self.contains(Key::MOTION_VECTOR_PREPASS) {
-                "_motion"
-            } else {
-                Default::default()
-            },
-            if self.contains(Key::DEFERRED_PREPASS) {
-                "_deferred"
-            } else {
-                Default::default()
-            },
-            if self.contains(Key::OIT_ENABLED) {
-                "_oit"
-            } else {
-                Default::default()
-            },
-        )
-    }
-}
-
-impl From<MeshPipelineKey> for MeshPipelineViewLayoutKey {
-    fn from(value: MeshPipelineKey) -> Self {
-        let mut result = MeshPipelineViewLayoutKey::empty();
-
-        if value.msaa_samples() > 1 {
-            result |= MeshPipelineViewLayoutKey::MULTISAMPLED;
-        }
-        if value.contains(MeshPipelineKey::DEPTH_PREPASS) {
-            result |= MeshPipelineViewLayoutKey::DEPTH_PREPASS;
-        }
-        if value.contains(MeshPipelineKey::NORMAL_PREPASS) {
-            result |= MeshPipelineViewLayoutKey::NORMAL_PREPASS;
-        }
-        if value.contains(MeshPipelineKey::MOTION_VECTOR_PREPASS) {
-            result |= MeshPipelineViewLayoutKey::MOTION_VECTOR_PREPASS;
-        }
-        if value.contains(MeshPipelineKey::DEFERRED_PREPASS) {
-            result |= MeshPipelineViewLayoutKey::DEFERRED_PREPASS;
-        }
-        if value.contains(MeshPipelineKey::OIT_ENABLED) {
-            result |= MeshPipelineViewLayoutKey::OIT_ENABLED;
-        }
-
-        result
-    }
-}
 
 pub fn mesh_pipeline_view_layout_key_from_msaa(value: Msaa) -> MeshPipelineViewLayoutKey {
     let mut result = MeshPipelineViewLayoutKey::empty();
@@ -412,15 +309,6 @@ fn layout_entries(
     [entries.to_vec(), binding_array_entries.to_vec()]
 }
 
-/// Stores the view layouts for every combination of pipeline keys.
-///
-/// This is wrapped in an [`Arc`] so that it can be efficiently cloned and
-/// placed inside specializable pipeline types.
-#[derive(Resource, Clone, Deref, DerefMut)]
-pub struct MeshPipelineViewLayouts(
-    pub Arc<[MeshPipelineViewLayout; MeshPipelineViewLayoutKey::COUNT]>,
-);
-
 pub fn init_mesh_pipeline_view_layouts(world: &mut World) {
     // Generates all possible view layouts for the mesh pipeline, based on all combinations of
     // [`MeshPipelineViewLayoutKey`] flags.
@@ -464,24 +352,6 @@ pub fn init_mesh_pipeline_view_layouts(world: &mut World) {
     })));
 
     world.insert_resource(res);
-}
-
-impl MeshPipelineViewLayouts {
-    pub fn get_view_layout(
-        &self,
-        layout_key: MeshPipelineViewLayoutKey,
-    ) -> &MeshPipelineViewLayout {
-        let index = layout_key.bits() as usize;
-        let layout = &self[index];
-
-        #[cfg(debug_assertions)]
-        if layout.texture_count > MESH_PIPELINE_VIEW_LAYOUT_SAFE_MAX_TEXTURES {
-            // Issue our own warning here because Naga's error message is a bit cryptic in this situation
-            once!(warn!("Too many textures in mesh pipeline view layout, this might cause us to hit `wgpu::Limits::max_sampled_textures_per_shader_stage` in some environments."));
-        }
-
-        layout
-    }
 }
 
 /// Generates all possible view layouts for the mesh pipeline, based on all combinations of
